extract some functions and consts to own file
